move code out of scope && fix issue
